adds tracklist uri array and refactors mock api response and mock api playlist to ExampleContent.js
